requirements.tests.txt
# adverse-event-app
django-celery-beat
django-celery-results
django-collect-offline
django-collect-offline-files
django-crypto-fields
edc-adverse-event
edc-appointment
edc-auth
edc-consent
edc-dashboard
edc-data-manager
edc-device
edc-export
edc-facility
edc-identifier
edc-lab
edc-lab-dashboard
edc-list-data
edc-locator
edc-metadata
edc-metadata-rules
edc-navbar
edc-notification
edc-offstudy
edc-pharmacy
edc-randomization
edc-reference
edc-registration
edc-review-dashboard
edc-sites
edc-subject-dashboard
edc-timepoint
edc-visit-schedule
edc-visit-tracking
